+++ src/Lik.cpp
@@ -33,7 +33,7 @@ double ObsLogLik(const arma::mat & mpar,
                   arma::mat vlogPrior,
                   arma::vec  vgroup,
                   arma::mat & mloc,
-                  arma::vec weights){
+                  arma::vec & weights){
 
 
   arma::mat mP = uP(mloc, mpar);
@@ -49,9 +49,9 @@ double ObsLogLik(const arma::mat & mpar,
     arma::mat mX1 = arma::ones<arma::mat>(arma::size(mX)); //missing -> 1
     mX0.elem( arma::find_nonfinite(mX) ).zeros(); //missing - 0
     mX1.elem( arma::find_nonfinite(mX) ).ones(); //missing - 1
-    mlogLik = mX0*log(mP) + (1-mX1)*log(1-mP); //N x L
+    mlogLik = mX0*arma::trunc_log(mP) + (1-mX1)*arma::trunc_log(1-mP); //N x L
   }else{
-    mlogLik = mX*log(mP) + (1-mX)*log(1-mP); //N x L
+    mlogLik = mX*arma::trunc_log(mP) + (1-mX)*arma::trunc_log(1-mP); //N x L
   }
 
   //joint prob (X_i, alpha_c) = log P(X_i|alpha_c) + log p(alpha_c)
@@ -66,7 +66,7 @@ double ObsLogLik(const arma::mat & mpar,
   }
 
   //sum_i weight_i log[ sum_c P(X_i|alpha_c)p(alpha_c)]
-  double LL = arma::accu(log(arma::sum(exp(mlogPost),1))%weights);
+  double LL = arma::accu(arma::trunc_log(arma::sum(arma::trunc_exp(mlogPost),1))%weights);
   return LL;
 }
 
@@ -81,9 +81,9 @@ Rcpp::List LikNR(const arma::mat & mpar,
                   arma::mat vlogPrior, //a vector of log prior or a matrix of log prior: col 1 for group 1; col 2 for group 2
                   arma::vec  vgroup,
                   arma::mat  mloc,
-                  arma::vec weights,
-                  int simplify = 1){
-
+                  arma::vec & weights,
+                  int simplify = true){
+  Rcpp::List ret;
   // calculate mP L x J matrix P(Xij=1|alpha_c)
 
   arma::mat mP = uP(mloc, mpar);
@@ -110,14 +110,14 @@ Rcpp::List LikNR(const arma::mat & mpar,
     mX1.elem( missingloc ).ones(); //missing - 1
     mXMissing.elem(missingloc).zeros();
     mXMissing.elem(arma::find_finite(mX)).ones();
-    mlogLik = mX0*log(mP) + (1-mX1)*log(1-mP); //N x L
+    mlogLik = mX0*arma::trunc_log(mP) + (1-mX1)*arma::trunc_log(1-mP); //N x L
   }else{
-    mlogLik = mX*log(mP) + (1-mX)*log(1-mP); //N x L
+    mlogLik = mX*arma::trunc_log(mP) + (1-mX)*arma::trunc_log(1-mP); //N x L
   }
 
   //joint prob (X_i, alpha_c) = log P(X_i|alpha_c) + log p(alpha_c)
   if(no_mg==1){
-    mlogPost = mlogLik.each_row() + vlogPrior.t(); // N x L
+    mlogPost = mlogLik + arma::ones<arma::mat>(N,1) * arma::trans(vlogPrior); // N x L
   }else{
     mlogPost = mlogLik;
     for (int g=0;g<no_mg;g++){
@@ -125,22 +125,32 @@ Rcpp::List LikNR(const arma::mat & mpar,
       mlogPost.rows(locg) += arma::ones<arma::mat>(locg.n_elem,1)*arma::trans(vlogPrior.col(g)); // N x L
     }
   }
-  arma::mat mPost = exp(mlogPost);// P(X_i|alpha_c)p(alpha_c)
-  double LL = arma::accu(log(arma::sum(mPost,1))%weights); //sum_i weight_i log[ sum_c P(X_i|alpha_c)p(alpha_c)]
+  arma::mat msdPost = arma::trunc_exp(mlogPost);// P(X_i|alpha_c)p(alpha_c)
+  arma::vec rowsum_msdPost = arma::sum(msdPost,1); //Nx1 sum_c P(X_i|alpha_c)p(alpha_c)
+  double LL = arma::dot(arma::trunc_log(rowsum_msdPost),weights); //sum_i weight_i log[ sum_c P(X_i|alpha_c)p(alpha_c)]
+  ret["LL"]=LL;
+
   // normalized posterior log [P(alpha_c|X_i)]
-  arma::mat msdPost = mPost;
-  msdPost.each_col()/=arma::sum(mPost,1); //N x L
+
+  msdPost/=rowsum_msdPost*arma::ones<arma::mat>(1,L); //N x L
+
+  if(simplify==false){
+    ret["loglik"]=mlogLik;
+    ret["logpost"]=arma::trunc_log(msdPost);
+  }
+  //-----> weighted msdPost
   msdPost.each_col() %= weights;
 
   arma::mat updlogPrior = arma::ones<arma::mat>(L,no_mg);
   if(no_mg==1){
-    updlogPrior = log(arma::trans(arma::sum(msdPost,0)/arma::accu(weights)));//L x 1
+    updlogPrior = arma::trunc_log(arma::trans(arma::sum(msdPost,0)/arma::accu(weights)));//L x 1
   }else{
     for (int g=0;g<no_mg;g++){
       locg = arma::find(vgroup==g);
-      updlogPrior.col(g) = log(arma::sum(msdPost.rows(locg),0).t()/arma::accu(weights(locg))); //updated log priors
+      updlogPrior.col(g) = arma::trunc_log(arma::sum(msdPost.rows(locg),0).t()/arma::accu(weights(locg))); //updated log priors
     }
   }
+  ret["logprior"]=updlogPrior;
 
 
   arma::mat Ng = arma::zeros<arma::mat>(J,Ljmax);
@@ -169,16 +179,9 @@ Rcpp::List LikNR(const arma::mat & mpar,
 
     }
 
-
-  Rcpp::List ret;
-  ret["LL"]=LL;
-  ret["logprior"]=updlogPrior;
   ret["Ng"]=Ng;
   ret["Rg"]=Rg;
-  if(simplify==0){
-    ret["loglik"]=mlogLik;
-    ret["logpost"]=log(msdPost);
-  }
+
 
   return ret;
 }