+++ src/RcppExports.cpp
@@ -19,7 +19,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // ObsLogLik
-double ObsLogLik(const arma::mat& mpar, const arma::mat& mX, arma::mat vlogPrior, arma::vec vgroup, arma::mat& mloc, arma::vec weights);
+double ObsLogLik(const arma::mat& mpar, const arma::mat& mX, arma::mat vlogPrior, arma::vec vgroup, arma::mat& mloc, arma::vec& weights);
 RcppExport SEXP _GDINA_ObsLogLik(SEXP mparSEXP, SEXP mXSEXP, SEXP vlogPriorSEXP, SEXP vgroupSEXP, SEXP mlocSEXP, SEXP weightsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -29,13 +29,13 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::mat >::type vlogPrior(vlogPriorSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type vgroup(vgroupSEXP);
     Rcpp::traits::input_parameter< arma::mat& >::type mloc(mlocSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type weights(weightsSEXP);
+    Rcpp::traits::input_parameter< arma::vec& >::type weights(weightsSEXP);
     rcpp_result_gen = Rcpp::wrap(ObsLogLik(mpar, mX, vlogPrior, vgroup, mloc, weights));
     return rcpp_result_gen;
 END_RCPP
 }
 // LikNR
-Rcpp::List LikNR(const arma::mat& mpar, const arma::mat& mX, arma::mat vlogPrior, arma::vec vgroup, arma::mat mloc, arma::vec weights, int simplify);
+Rcpp::List LikNR(const arma::mat& mpar, const arma::mat& mX, arma::mat vlogPrior, arma::vec vgroup, arma::mat mloc, arma::vec& weights, int simplify);
 RcppExport SEXP _GDINA_LikNR(SEXP mparSEXP, SEXP mXSEXP, SEXP vlogPriorSEXP, SEXP vgroupSEXP, SEXP mlocSEXP, SEXP weightsSEXP, SEXP simplifySEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -45,7 +45,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::mat >::type vlogPrior(vlogPriorSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type vgroup(vgroupSEXP);
     Rcpp::traits::input_parameter< arma::mat >::type mloc(mlocSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type weights(weightsSEXP);
+    Rcpp::traits::input_parameter< arma::vec& >::type weights(weightsSEXP);
     Rcpp::traits::input_parameter< int >::type simplify(simplifySEXP);
     rcpp_result_gen = Rcpp::wrap(LikNR(mpar, mX, vlogPrior, vgroup, mloc, weights, simplify));
     return rcpp_result_gen;